+++ benchmarks/test_pydantic.py
@@ -10,9 +10,9 @@ class TestPydantic:
     def __init__(self, allow_extra):
 
         class Model(BaseModel):
-            id: int = ...
-            client_name: constr(max_length=255) = ...
-            sort_index: float = ...
+            id: int
+            client_name: constr(max_length=255)
+            sort_index: float
             # client_email: EmailStr = None
             client_phone: constr(max_length=255) = None
 
@@ -23,15 +23,15 @@ class TestPydantic:
 
             contractor: PositiveInt = None
             upstream_http_referrer: constr(max_length=1023) = None
-            grecaptcha_response: constr(min_length=20, max_length=1000) = ...
+            grecaptcha_response: constr(min_length=20, max_length=1000)
             last_updated: datetime = None
 
             class Skill(BaseModel):
-                subject: str = ...
-                subject_id: int = ...
-                category: str = ...
-                qual_level: str = ...
-                qual_level_id: int = ...
+                subject: str
+                subject_id: int
+                category: str
+                qual_level: str
+                qual_level_id: int
                 qual_level_ranking: float = 0
             skills: List[Skill] = []